feat: turn BLE code into hook
